requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
requests>=2.31.0
beautifulsoup4>=4.12.2
pdfplumber>=0.10.3
python-docx>=1.0.1
python-multipart>=0.0.6
pydantic>=2.6.0
python-jose>=3.3.0
passlib>=1.7.4
bcrypt>=4.0.1
python-dotenv>=1.0.0
aiofiles>=23.2.1
jinja2>=3.1.2
slowapi>=0.1.8
redis>=5.0.1
httpx>=0.25.2
tenacity>=8.2.3
openai>=1.3.7
nltk>=3.8.1